ai_learn/langchain_a/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
dashscope>=1.10.0
python-dotenv>=1.0.0
requests>=2.31.0
